requirements.txt
anthropic>=0.40.0
google-genai>=0.3.0
openpyxl>=3.1.2
pydantic>=2.6.0
rapidfuzz>=3.6.0
python-dotenv>=1.0.1
tenacity>=8.2.3
rich>=13.7.0
pyyaml>=6.0.1
